Login Page tutorial project
This is the project from https://www.youtube.com/watch?v=xb8aad4MRx8 . It contains a lot of good info on sessions, connecting to and getting info from a database. We c can use this stuff for the login part of the database. There is an sql file for creating the database too. I'll add more to the database - it just has one table for the users at the moment.